+++ test/tree_test.dart
@@ -350,14 +350,17 @@ void main() {
           messages = ((e as dynamic).message as String).split('\n');
         }
 
-        expect(messages, [
-          'Could not find node "unknown"',
-          '',
-          'Possible paths:',
-          '  - .',
-          '  - ./child',
-          '  - ./child/grandchild',
-        ]);
+        expect(
+          messages,
+          containsAll([
+            'Could not find node "unknown"',
+            '',
+            'Possible paths:',
+            '  - .',
+            '  - ./child',
+            '  - ./child/grandchild',
+          ]),
+        );
       });
     });
 
@@ -383,14 +386,17 @@ void main() {
               messages = ((e as dynamic).message as String).split('\n');
             }
 
-            expect(messages, [
-              'Could not find node "/grandpa/dad/me/unknown"',
-              '',
-              'Possible paths:',
-              '  - /grandpa/dad/me',
-              '  - /grandpa/dad/me/child',
-              '  - /grandpa/dad/me/child/grandchild',
-            ]);
+            expect(
+              messages,
+              containsAll([
+                'Could not find node "/grandpa/dad/me/unknown"',
+                '',
+                'Possible paths:',
+                '  - /grandpa/dad/me',
+                '  - /grandpa/dad/me/child',
+                '  - /grandpa/dad/me/child/grandchild',
+              ]),
+            );
           });
 
           test('and an ancestor node is not found', () {
@@ -401,14 +407,17 @@ void main() {
               message = (e as dynamic).message.toString().trim().split('\n');
             }
 
-            expect(message, [
-              'Could not find node "/grandpa/dad/unknown"',
-              '',
-              'Possible paths:',
-              '  - /grandpa/dad',
-              '  - /grandpa/dad/child',
-              '  - /grandpa/dad/child/grandchild',
-            ]);
+            expect(
+              message,
+              containsAll([
+                'Could not find node "/grandpa/dad/unknown"',
+                '',
+                'Possible paths:',
+                '  - /grandpa/dad',
+                '  - /grandpa/dad/child',
+                '  - /grandpa/dad/child/grandchild',
+              ]),
+            );
           });
         });
       });
@@ -429,14 +438,17 @@ void main() {
             messages = ((e as dynamic).message as String).split('\n');
           }
 
-          expect(messages, [
-            'Could not find node "../me/unknown"',
-            '',
-            'Possible paths:',
-            '  - ../me',
-            '  - ../me/child',
-            '  - ../me/child/grandchild',
-          ]);
+          expect(
+            messages,
+            containsAll([
+              'Could not find node "../me/unknown"',
+              '',
+              'Possible paths:',
+              '  - ../me',
+              '  - ../me/child',
+              '  - ../me/child/grandchild',
+            ]),
+          );
         });
       });
     });
@@ -560,6 +572,127 @@ void main() {
             '  - ./nums/false',
             '  - ./nums/null',
           ]);
+        });
+      });
+
+      group('returns data from', () {
+        group('nodes directly', () {
+          test('using ./', () {
+            expect(me.get<String>('./#hiMe'), 'Hi from me.');
+
+            expect(me.get<int>('./#nums/int'), 42);
+            expect(me.get<double>('./#nums/double'), 3.14159);
+            expect(me.get<bool>('./#nums/true'), isTrue);
+            expect(me.get<bool>('./#nums/false'), isFalse);
+          });
+
+          group('parent node', () {
+            test('using ../', () {
+              expect(me.get<String>('../#hiDad'), 'Hi from dad.');
+
+              expect(me.get<int>('../#nums/int'), 42);
+              expect(me.get<double>('../#nums/double'), 3.14159);
+              expect(me.get<bool>('../#nums/true'), isTrue);
+              expect(me.get<bool>('../#nums/false'), isFalse);
+            });
+          });
+
+          group('self or any ancestors', () {
+            test('using no operator', () {
+              expect(me.get<String>('#hiRoot'), 'Hi from root.');
+              expect(me.get<String>('#hiGrandpa'), 'Hi from grandpa.');
+              expect(me.get<String>('#hiDad'), 'Hi from dad.');
+              expect(me.get<String>('#hiMe'), 'Hi from me.');
+
+              expect(me.get<int>('#nums/int'), 42);
+              expect(me.get<double>('#nums/double'), 3.14159);
+              expect(me.get<bool>('#nums/true'), isTrue);
+              expect(me.get<bool>('#nums/false'), isFalse);
+            });
+          });
+        });
+
+        group('child nodes', () {
+          group('i.e. from own child nodes', () {
+            test('using ./child', () {
+              expect(me.get<String>('./child#hiChild'), 'Hi from child.');
+
+              expect(me.get<int>('./child#nums/int'), 42);
+              expect(me.get<double>('./child#nums/double'), 3.14159);
+              expect(me.get<bool>('./child#nums/true'), isTrue);
+              expect(me.get<bool>('./child#nums/false'), isFalse);
+            });
+
+            test('using ./child/grandchild', () {
+              expect(
+                me.get<String>('./child/grandchild#hiGrandchild'),
+                'Hi from grandchild.',
+              );
+
+              expect(me.get<int>('./child/grandchild#nums/int'), 42);
+              expect(me.get<double>('./child/grandchild#nums/double'), 3.14159);
+              expect(me.get<bool>('./child/grandchild#nums/true'), isTrue);
+              expect(me.get<bool>('./child/grandchild#nums/false'), isFalse);
+            });
+          });
+        });
+
+        group('the node itself', () {
+          group('using node', () {
+            test('node/key', () {
+              expect(me.get<String>('./#node/key'), 'me');
+            });
+            test('node/childCount', () {
+              expect(me.get<int>('./#node/childCount'), 1);
+              expect(dad.get<int>('./#node/childCount'), 3);
+            });
+            test('node/index', () {
+              expect(me.get<int>('./#node/index'), 0);
+              expect(brother.get<int>('./#node/index'), 1);
+              expect(sister.get<int>('./#node/index'), 2);
+              expect(root.get<int>('./#node/index'), 0);
+            });
+            test('node/siblingsCount', () {
+              expect(me.get<int>('./#node/siblingsCount'), 3);
+              expect(brother.get<int>('./#node/siblingsCount'), 3);
+              expect(sister.get<int>('./#node/siblingsCount'), 3);
+              expect(root.get<int>('./#node/siblingsCount'), 1);
+            });
+            test('node/reverseIndex', () {
+              expect(me.get<int>('./#node/reverseIndex'), 2);
+              expect(brother.get<int>('./#node/reverseIndex'), 1);
+              expect(sister.get<int>('./#node/reverseIndex'), 0);
+              expect(root.get<int>('./#node/reverseIndex'), 0);
+            });
+            test('node/path', () {
+              expect(me.get<String>('./#node/path'), '/grandpa/dad/me');
+              expect(
+                brother.get<String>('./#node/path'),
+                '/grandpa/dad/brother',
+              );
+              expect(sister.get<String>('./#node/path'), '/grandpa/dad/sister');
+              expect(root.get<String>('./#node/path'), '/');
+            });
+            test('node/pathSimple', () {
+              me.key = 'me89';
+              brother.key = 'brother_90';
+
+              expect(me.get<String>('./#node/pathSimple'), '/grandpa/dad/me');
+              expect(
+                brother.get<String>('./#node/pathSimple'),
+                '/grandpa/dad/brother',
+              );
+              expect(
+                sister.get<String>('./#node/pathSimple'),
+                '/grandpa/dad/sister',
+              );
+              expect(root.get<String>('./#node/pathSimple'), '/');
+            });
+            test('node/isRoot', () {
+              expect(root.get<bool>('./#node/isRoot'), isTrue);
+              expect(sister.get<bool>('./#node/isRoot'), isFalse);
+            });
+          });
         });
       });
     });
@@ -879,32 +1012,287 @@ void main() {
     });
 
     group('ls', () {
-      test('return all paths of the tree', () {
-        expect(root.ls(), [
-          '.',
-          './grandpa',
-          './grandpa/dad',
-          './grandpa/dad/me',
-          './grandpa/dad/me/child',
-          './grandpa/dad/me/child/grandchild',
-          './grandpa/dad/brother',
-          './grandpa/dad/sister',
-        ]);
+      group('with showDataPaths', () {
+        group('true', () {
+          test('return all paths of the tree', () async {
+            final ls = root.ls();
+            await writeGolden('root_ls.json', ls);
 
-        expect(me.ls(), ['.', './child', './child/grandchild']);
+            expect(
+              root.ls(),
+              containsAll([
+                '.',
+                '.#me',
+                '.#hiRoot',
+                '.#isAncestor',
+                '.#node/key',
+                '.#node/originalKey',
+                '.#node/isReadOnly',
+                '.#node/childCount',
+                '.#node/siblingsCount',
+                '.#node/index',
+                '.#node/reverseIndex',
+                '.#node/path',
+                '.#node/pathSimple',
+                '.#node/isRoot',
+                './grandpa',
+                './grandpa#me',
+                './grandpa#hiGrandpa',
+                './grandpa#isAncestor',
+                './grandpa#node/key',
+                './grandpa#node/originalKey',
+                './grandpa#node/isReadOnly',
+                './grandpa#node/childCount',
+                './grandpa#node/siblingsCount',
+                './grandpa#node/index',
+                './grandpa#node/reverseIndex',
+                './grandpa#node/path',
+                './grandpa#node/pathSimple',
+                './grandpa#node/isRoot',
+                './grandpa/dad',
+                './grandpa/dad#me',
+                './grandpa/dad#hiDad',
+                './grandpa/dad#isAncestor',
+                './grandpa/dad#nums',
+                './grandpa/dad#nums/string',
+                './grandpa/dad#nums/int',
+                './grandpa/dad#nums/double',
+                './grandpa/dad#nums/true',
+                './grandpa/dad#nums/false',
+                './grandpa/dad#nums/null',
+                './grandpa/dad#node/key',
+                './grandpa/dad#node/originalKey',
+                './grandpa/dad#node/isReadOnly',
+                './grandpa/dad#node/childCount',
+                './grandpa/dad#node/siblingsCount',
+                './grandpa/dad#node/index',
+                './grandpa/dad#node/reverseIndex',
+                './grandpa/dad#node/path',
+                './grandpa/dad#node/pathSimple',
+                './grandpa/dad#node/isRoot',
+                './grandpa/dad/me',
+                './grandpa/dad/me#me',
+                './grandpa/dad/me#hiMe',
+                './grandpa/dad/me#isAncestor',
+                './grandpa/dad/me#nums',
+                './grandpa/dad/me#nums/string',
+                './grandpa/dad/me#nums/int',
+                './grandpa/dad/me#nums/double',
+                './grandpa/dad/me#nums/true',
+                './grandpa/dad/me#nums/false',
+                './grandpa/dad/me#nums/null',
+                './grandpa/dad/me#node/key',
+                './grandpa/dad/me#node/originalKey',
+                './grandpa/dad/me#node/isReadOnly',
+                './grandpa/dad/me#node/childCount',
+                './grandpa/dad/me#node/siblingsCount',
+                './grandpa/dad/me#node/index',
+                './grandpa/dad/me#node/reverseIndex',
+                './grandpa/dad/me#node/path',
+                './grandpa/dad/me#node/pathSimple',
+                './grandpa/dad/me#node/isRoot',
+                './grandpa/dad/me/child',
+                './grandpa/dad/me/child#me',
+                './grandpa/dad/me/child#hiChild',
+                './grandpa/dad/me/child#isAncestor',
+                './grandpa/dad/me/child#nums',
+                './grandpa/dad/me/child#nums/string',
+                './grandpa/dad/me/child#nums/int',
+                './grandpa/dad/me/child#nums/double',
+                './grandpa/dad/me/child#nums/true',
+                './grandpa/dad/me/child#nums/false',
+                './grandpa/dad/me/child#nums/null',
+                './grandpa/dad/me/child#node/key',
+                './grandpa/dad/me/child#node/originalKey',
+                './grandpa/dad/me/child#node/isReadOnly',
+                './grandpa/dad/me/child#node/childCount',
+                './grandpa/dad/me/child#node/siblingsCount',
+                './grandpa/dad/me/child#node/index',
+                './grandpa/dad/me/child#node/reverseIndex',
+                './grandpa/dad/me/child#node/path',
+                './grandpa/dad/me/child#node/pathSimple',
+                './grandpa/dad/me/child#node/isRoot',
+                './grandpa/dad/me/child/grandchild',
+                './grandpa/dad/me/child/grandchild#me',
+                './grandpa/dad/me/child/grandchild#hiGrandchild',
+                './grandpa/dad/me/child/grandchild#isAncestor',
+                './grandpa/dad/me/child/grandchild#nums',
+                './grandpa/dad/me/child/grandchild#nums/string',
+                './grandpa/dad/me/child/grandchild#nums/int',
+                './grandpa/dad/me/child/grandchild#nums/double',
+                './grandpa/dad/me/child/grandchild#nums/true',
+                './grandpa/dad/me/child/grandchild#nums/false',
+                './grandpa/dad/me/child/grandchild#nums/null',
+                './grandpa/dad/me/child/grandchild#node/key',
+                './grandpa/dad/me/child/grandchild#node/originalKey',
+                './grandpa/dad/me/child/grandchild#node/isReadOnly',
+                './grandpa/dad/me/child/grandchild#node/childCount',
+                './grandpa/dad/me/child/grandchild#node/siblingsCount',
+                './grandpa/dad/me/child/grandchild#node/index',
+                './grandpa/dad/me/child/grandchild#node/reverseIndex',
+                './grandpa/dad/me/child/grandchild#node/path',
+                './grandpa/dad/me/child/grandchild#node/pathSimple',
+                './grandpa/dad/me/child/grandchild#node/isRoot',
+                './grandpa/dad/brother',
+                './grandpa/dad/brother#me',
+                './grandpa/dad/brother#hiBrother',
+                './grandpa/dad/brother#isAncestor',
+                './grandpa/dad/brother#nums',
+                './grandpa/dad/brother#nums/string',
+                './grandpa/dad/brother#nums/int',
+                './grandpa/dad/brother#nums/double',
+                './grandpa/dad/brother#nums/true',
+                './grandpa/dad/brother#nums/false',
+                './grandpa/dad/brother#nums/null',
+                './grandpa/dad/brother#node/key',
+                './grandpa/dad/brother#node/originalKey',
+                './grandpa/dad/brother#node/isReadOnly',
+                './grandpa/dad/brother#node/childCount',
+                './grandpa/dad/brother#node/siblingsCount',
+                './grandpa/dad/brother#node/index',
+                './grandpa/dad/brother#node/reverseIndex',
+                './grandpa/dad/brother#node/path',
+                './grandpa/dad/brother#node/pathSimple',
+                './grandpa/dad/brother#node/isRoot',
+                './grandpa/dad/sister',
+                './grandpa/dad/sister#me',
+                './grandpa/dad/sister#hiBrother',
+                './grandpa/dad/sister#isAncestor',
+                './grandpa/dad/sister#nums',
+                './grandpa/dad/sister#nums/string',
+                './grandpa/dad/sister#nums/int',
+                './grandpa/dad/sister#nums/double',
+                './grandpa/dad/sister#nums/true',
+                './grandpa/dad/sister#nums/false',
+                './grandpa/dad/sister#nums/null',
+                './grandpa/dad/sister#node/key',
+                './grandpa/dad/sister#node/originalKey',
+                './grandpa/dad/sister#node/isReadOnly',
+                './grandpa/dad/sister#node/childCount',
+                './grandpa/dad/sister#node/siblingsCount',
+                './grandpa/dad/sister#node/index',
+                './grandpa/dad/sister#node/reverseIndex',
+                './grandpa/dad/sister#node/path',
+                './grandpa/dad/sister#node/pathSimple',
+                './grandpa/dad/sister#node/isRoot',
+              ]),
+            );
+
+            final meLs = me.ls();
+            await writeGolden('me_ls.json', meLs);
+
+            expect(meLs, [
+              '.',
+              '.#me',
+              '.#hiMe',
+              '.#isAncestor',
+              '.#nums',
+              '.#nums/string',
+              '.#nums/int',
+              '.#nums/double',
+              '.#nums/true',
+              '.#nums/false',
+              '.#nums/null',
+              '.#node/key',
+              '.#node/originalKey',
+              '.#node/isReadOnly',
+              '.#node/childCount',
+              '.#node/siblingsCount',
+              '.#node/index',
+              '.#node/reverseIndex',
+              '.#node/path',
+              '.#node/pathSimple',
+              '.#node/isRoot',
+              './child',
+              './child#me',
+              './child#hiChild',
+              './child#isAncestor',
+              './child#nums',
+              './child#nums/string',
+              './child#nums/int',
+              './child#nums/double',
+              './child#nums/true',
+              './child#nums/false',
+              './child#nums/null',
+              './child#node/key',
+              './child#node/originalKey',
+              './child#node/isReadOnly',
+              './child#node/childCount',
+              './child#node/siblingsCount',
+              './child#node/index',
+              './child#node/reverseIndex',
+              './child#node/path',
+              './child#node/pathSimple',
+              './child#node/isRoot',
+              './child/grandchild',
+              './child/grandchild#me',
+              './child/grandchild#hiGrandchild',
+              './child/grandchild#isAncestor',
+              './child/grandchild#nums',
+              './child/grandchild#nums/string',
+              './child/grandchild#nums/int',
+              './child/grandchild#nums/double',
+              './child/grandchild#nums/true',
+              './child/grandchild#nums/false',
+              './child/grandchild#nums/null',
+              './child/grandchild#node/key',
+              './child/grandchild#node/originalKey',
+              './child/grandchild#node/isReadOnly',
+              './child/grandchild#node/childCount',
+              './child/grandchild#node/siblingsCount',
+              './child/grandchild#node/index',
+              './child/grandchild#node/reverseIndex',
+              './child/grandchild#node/path',
+              './child/grandchild#node/pathSimple',
+              './child/grandchild#node/isRoot',
+            ]);
+          });
+        });
+
+        group('false', () {
+          test('shows only the node paths', () async {
+            final ls = root.ls(showDataPaths: false);
+            await writeGolden('root_ls_without_data_paths.json', ls);
+
+            expect(
+              ls,
+              containsAll([
+                '.',
+                './grandpa',
+                './grandpa/dad',
+                './grandpa/dad/me',
+                './grandpa/dad/me/child',
+                './grandpa/dad/me/child/grandchild',
+                './grandpa/dad/brother',
+                './grandpa/dad/sister',
+              ]),
+            );
+
+            final meLs = me.ls(showDataPaths: false);
+            await writeGolden('me_ls.json', meLs);
+
+            expect(meLs, ['.', './child', './child/grandchild']);
+          });
+        });
       });
 
       group('with where', () {
         test('returns all paths matching the given condition', () {
-          expect(root.ls(where: (node) => node.key.startsWith('g')), [
-            './grandpa',
-            './grandpa/dad/me/child/grandchild',
-          ]);
+          expect(
+            root.ls(
+              where: (node) => node.key.startsWith('g'),
+              showDataPaths: false,
+            ),
+            ['./grandpa', './grandpa/dad/me/child/grandchild'],
+          );
 
-          expect(me.ls(where: (node) => node.key.contains('h')), [
-            './child',
-            './child/grandchild',
-          ]);
+          expect(
+            me.ls(
+              where: (node) => node.key.contains('h'),
+              showDataPaths: false,
+            ),
+            ['./child', './child/grandchild'],
+          );
         });
       });
     });